refactored vault management page
