+++ pages/manage/vaults/[id].tsx
@@ -1,63 +1,22 @@
-import AssetWithName from "@/components/vault/AssetWithName";
 import { vaultsAtom } from "@/lib/atoms/vaults";
-import {
-  VaultFees,
-  VaultData,
-  Token,
-} from "@/lib/types";
+import { VaultData, Token } from "@/lib/types";
 import { useAtom } from "jotai";
 import { useRouter } from "next/router";
-import {
-  Fragment,
-  useEffect,
-  useState,
-} from "react";
+import { useEffect, useState } from "react";
 import NoSSR from "react-no-ssr";
-import { useAccount, useNetwork, usePublicClient, useSwitchNetwork, useWalletClient } from "wagmi";
-import {
-  Address,
-  createPublicClient,
-  extractChain,
-  formatUnits,
-  http,
-  isAddress,
-} from "viem";
-import {
-  AdminProxyByChain,
-  MultiStrategyVaultAbi,
-  VaultAbi,
-  VaultControllerByChain,
-} from "@/lib/constants";
+import { createPublicClient, http, isAddress } from "viem";
+import { VaultAbi } from "@/lib/constants";
 import { ChainById, RPC_URLS } from "@/lib/utils/connectors";
-import * as chains from "viem/chains";
-import { YieldOptions } from "vaultcraft-sdk";
 import { yieldOptionsAtom } from "@/lib/atoms/sdk";
-import TabSelector from "@/components/common/TabSelector";
-import VaultStrategyConfiguration from "@/components/vault/management/vault/Strategy";
-import VaultPausing from "@/components/vault/management/vault/Pausing";
-import VaultDepositLimit from "@/components/vault/management/vault/DepositLimit";
-import VaultFeeRecipient from "@/components/vault/management/vault/FeeRecipient";
-import VaultFeeConfiguration from "@/components/vault/management/vault/FeeConfiguration";
-import VaultTakeFees from "@/components/vault/management/vault/Fees";
-import VaultStrategiesConfiguration from "@/components/vault/management/vault/Strategies";
-import VaultRebalance from "@/components/vault/management/vault/Rebalance";
 import { tokensAtom } from "@/lib/atoms";
-import SecondaryActionButton from "@/components/button/SecondaryActionButton";
-import MainActionButton from "@/components/button/MainActionButton";
-import { formatAndRoundNumber, safeRound } from "@/lib/utils/formatBigNumber";
-import { validateInput, } from "@/lib/utils/helpers";
 import { showSuccessToast } from "@/lib/toasts";
 import CopyToClipboard from "react-copy-to-clipboard";
 import { Square2StackIcon } from "@heroicons/react/24/outline";
 import LeftArrowIcon from "@/components/svg/LeftArrowIcon";
 import ApyChart from "@/components/vault/management/vault/ApyChart";
 import NetFlowChart from "@/components/vault/management/vault/NetFlowChart";
-import TokenIcon from "@/components/common/TokenIcon";
-import { getRewardData } from "@/lib/gauges/useGaugeRewardData";
-import InputTokenWithError from "@/components/input/InputTokenWithError";
-import { fundReward } from "@/lib/gauges/interactions";
-import { handleAllowance } from "@/lib/approve";
 import VaultHero from "@/components/vault/VaultHero";
+import RewardsSection from "@/components/vault/management/vault/RewardsSection";
 
 async function getLogs(vault: VaultData, asset: Token) {
   const client = createPublicClient({
@@ -141,132 +100,11 @@ async function getLogs(vault: VaultData, asset: Token) {
   return result;
 }
 
-export interface VaultSettings {
-  proposedStrategies: Address[];
-  proposedStrategyTime: number;
-  proposedFees: VaultFees;
-  proposedFeeTime: number;
-  paused: boolean;
-  feeBalance: number;
-  accruedFees: number;
-  owner: Address;
-}
-
-function getMulticalls(vault: VaultData) {
-  const vaultContract = {
-    address: vault.address,
-    abi: vault.strategies.length > 1 ? MultiStrategyVaultAbi : VaultAbi,
-  };
-  const result = [
-    {
-      ...vaultContract,
-      functionName: "proposedFees",
-    },
-    {
-      ...vaultContract,
-      functionName: "proposedFeeTime",
-    },
-    {
-      ...vaultContract,
-      functionName: "paused",
-    },
-    {
-      ...vaultContract,
-      functionName: "balanceOf",
-      args: [vault.metadata.feeRecipient],
-    },
-    {
-      ...vaultContract,
-      functionName: "accruedManagementFee",
-    },
-    {
-      ...vaultContract,
-      functionName: "accruedPerformanceFee",
-    },
-    {
-      ...vaultContract,
-      functionName: "owner",
-    },
-  ];
-
-  if (vault.strategies.length > 1) {
-    return [
-      {
-        ...vaultContract,
-        functionName: "getProposedStrategies",
-      },
-      {
-        ...vaultContract,
-        functionName: "proposedStrategyTime",
-      },
-      ...result,
-    ];
-  } else {
-    return [
-      {
-        ...vaultContract,
-        functionName: "proposedAdapter",
-      },
-      {
-        ...vaultContract,
-        functionName: "proposedAdapterTime",
-      },
-      ...result,
-    ];
-  }
-}
-
-async function getVaultSettings(
-  vault: VaultData,
-  yieldOptions: YieldOptions
-): Promise<VaultSettings> {
-  const client = createPublicClient({
-    chain: extractChain({
-      chains: Object.values(chains),
-      // @ts-ignore
-      id: vault.chainId,
-    }),
-    transport: http(RPC_URLS[vault.chainId]),
-  });
-
-  const res: any[] = await client.multicall({
-    contracts: getMulticalls(vault),
-    allowFailure: false,
-  });
-
-  return {
-    proposedStrategies: vault.strategies.length > 1 ? res[0] : [res[0]],
-    proposedStrategyTime: Number(res[1]),
-    proposedFees: {
-      deposit: Number(res[2][0]),
-      withdrawal: Number(res[2][1]),
-      management: Number(res[2][2]),
-      performance: Number(res[2][3]),
-    },
-    proposedFeeTime: Number(res[3]),
-    paused: res[4],
-    feeBalance: Number(res[5]),
-    accruedFees: Number(res[6]) + Number(res[7]),
-    owner: res[8],
-  };
-}
-
-const DEFAULT_TABS = [
-  "Strategy",
-  "Fee Configuration",
-  "Fee Recipient",
-  "Take Fees",
-  "Deposit Limit",
-  "Pausing",
-];
-
 export default function Index() {
   const router = useRouter();
   const { query } = router;
 
   const [yieldOptions] = useAtom(yieldOptionsAtom);
-
-  const { address: account } = useAccount();
 
   const [vaults, setVaults] = useAtom(vaultsAtom);
   const [tokens] = useAtom(tokensAtom);
@@ -276,12 +114,8 @@ export default function Index() {
   const [vault, setVault] = useState<Token>();
 
   const [vaultData, setVaultData] = useState<VaultData>();
-  const [settings, setSettings] = useState<VaultSettings>();
-  const [callAddress, setCallAddress] = useState<Address>();
 
   const [logs, setLogs] = useState<any[]>([]);
-  const [availableTabs, setAvailableTabs] = useState<string[]>(DEFAULT_TABS);
-  const [tab, setTab] = useState<string>("Strategy");
 
   useEffect(() => {
     async function setupVault() {
@@ -294,25 +128,12 @@ export default function Index() {
           vault_,
           tokens[vault_.chainId][vault_.asset]
         );
-        const settings_ = await getVaultSettings(vault_, yieldOptions!);
-
         setAsset(tokens[vault_.chainId][vault_.asset]);
         setVault(tokens[vault_.chainId][vault_.vault]);
         if (vault_.gauge) setGauge(tokens[vault_.chainId][vault_.gauge]);
 
         setVaultData(vault_);
-        setSettings(settings_);
-        setCallAddress(
-          settings_.owner === AdminProxyByChain[vault_.chainId]
-            ? VaultControllerByChain[vault_.chainId]
-            : vault_.address
-        );
         setLogs(logs_);
-
-        if (vault_.strategies.length > 1) {
-          setAvailableTabs(["Rebalance", ...DEFAULT_TABS]);
-          setTab("Rebalance");
-        }
       }
     }
     if (
@@ -324,10 +145,6 @@ export default function Index() {
     )
       setupVault();
   }, [vaults, tokens, query, vaultData, yieldOptions]);
-
-  function changeTab(tab: string) {
-    setTab(tab);
-  }
 
   return (
     <NoSSR>
@@ -364,89 +181,6 @@ export default function Index() {
             </section>
           )}
 
-          <section className="md:border-b border-customNeutral100 py-10 px-4 md:px-8 text-white">
-            <h2 className="text-white font-bold text-2xl">Vault Settings</h2>
-            {account === vaultData.metadata.creator ? (
-              <>
-                <TabSelector
-                  className="mt-6 mb-12"
-                  availableTabs={availableTabs}
-                  activeTab={tab}
-                  setActiveTab={changeTab}
-                />
-                {settings && callAddress ? (
-                  <div>
-                    {tab === "Rebalance" && (
-                      <VaultRebalance
-                        vaultData={vaultData}
-                        settings={settings}
-                        callAddress={callAddress}
-                      />
-                    )}
-                    {tab === "Strategy" && (
-                      <>
-                        {vaultData.strategies.length > 1 ? (
-                          <VaultStrategiesConfiguration
-                            vaultData={vaultData}
-                            settings={settings}
-                            callAddress={callAddress}
-                          />
-                        ) : (
-                          <VaultStrategyConfiguration
-                            vaultData={vaultData}
-                            settings={settings}
-                            callAddress={callAddress}
-                          />
-                        )}
-                      </>
-                    )}
-                    {tab === "Fee Configuration" && (
-                      <VaultFeeConfiguration
-                        vaultData={vaultData}
-                        settings={settings}
-                        callAddress={callAddress}
-                      />
-                    )}
-                    {tab === "Fee Recipient" && (
-                      <VaultFeeRecipient
-                        vaultData={vaultData}
-                        settings={settings}
-                        callAddress={callAddress}
-                      />
-                    )}
-                    {tab === "Deposit Limit" && (
-                      <VaultDepositLimit
-                        vaultData={vaultData}
-                        settings={settings}
-                        callAddress={callAddress}
-                      />
-                    )}
-                    {tab === "Take Fees" && (
-                      <VaultTakeFees
-                        vaultData={vaultData}
-                        settings={settings}
-                        callAddress={callAddress}
-                      />
-                    )}
-                    {tab === "Pausing" && (
-                      <VaultPausing
-                        vaultData={vaultData}
-                        settings={settings}
-                        callAddress={callAddress}
-                      />
-                    )}
-                  </div>
-                ) : (
-                  <p className="text-white">Loading...</p>
-                )}
-              </>
-            ) : (
-              <p className="text-white">
-                Only the Vault Creator ({vaultData.metadata.creator}) has access
-                to this page.
-              </p>
-            )}
-          </section>
 
           <section className="md:border-b border-customNeutral100 py-10 px-4 md:px-8 text-white">
             <div className="grid md:grid-cols-2 mb-12">
@@ -518,169 +252,4 @@ export default function Index() {
       )}
     </NoSSR>
   );
-}
-
-function RewardsSection({ gauge, chainId }: { gauge: Address; chainId: number }) {
-  const [tokens] = useAtom(tokensAtom);
-
-  const [rewardData, setRewardData] = useState<any[]>([])
-
-  useEffect(() => {
-    getRewardData(gauge, chainId).then(res => setRewardData(res))
-  }, [gauge])
-
-  return (
-    <Fragment>
-      <table className="mb-8 [&_th]:px-6 [&_th]:h-12 [&_td]:px-6 [&_td]:h-12 w-full">
-        <tr className="border-b">
-          <th className="font-normal text-left !pl-0">Token</th>
-          <th className="font-normal text-left">Rate</th>
-          <th className="font-normal text-left">Remaining Rewards</th>
-          <th className="font-normal text-left">Period Finish</th>
-          <th className="font-normal text-left hidden lg:table-cell">
-            Distributor
-          </th>
-          <th className="font-normal text-left hidden lg:table-cell">
-            Fund Rewards
-          </th>
-        </tr>
-        {rewardData.length > 0 ?
-          rewardData?.map(reward =>
-            <RewardColumn
-              key={`reward-gauge-${reward.address}`}
-              gauge={gauge}
-              reward={reward}
-              token={tokens[chainId][reward.address]}
-              chainId={chainId}
-            />
-          )
-          : <p>No rewards</p>
-        }
-      </table>
-    </Fragment>
-  );
-}
-
-
-function RewardColumn({ gauge, reward, token, chainId }: { gauge: Address, reward: any, token: Token, chainId: number }): JSX.Element {
-  const { address: account } = useAccount();
-  const publicClient = usePublicClient({ chainId });
-  const { data: walletClient } = useWalletClient();
-  const { chain } = useNetwork();
-  const { switchNetworkAsync } = useSwitchNetwork();
-  const [tokens, setTokens] = useAtom(tokensAtom);
-
-  const [amount, setAmount] = useState<string>("0");
-
-  function handleChangeInput(e: any) {
-    const value = e.currentTarget.value;
-    setAmount(validateInput(value).isValid ? value : "0");
-  }
-
-  function handleMaxClick() {
-    if (!token) return;
-    const stringBal = token.balance.toLocaleString("fullwide", {
-      useGrouping: false,
-    });
-    const rounded = safeRound(BigInt(stringBal), token.decimals);
-    const formatted = formatUnits(rounded, token.decimals);
-    handleChangeInput({ currentTarget: { value: formatted } });
-  }
-
-  async function handleApprove() {
-    let val = Number(amount)
-    if (val === 0 || !account || !publicClient || !walletClient) return
-    val = val * (10 ** token.decimals)
-
-    if (chain?.id !== Number(chainId)) {
-      try {
-        await switchNetworkAsync?.(Number(chainId));
-      } catch (error) {
-        return;
-      }
-    }
-
-    handleAllowance({
-      token: token.address,
-      spender: gauge,
-      amount: val,
-      account: account,
-      clients: {
-        publicClient,
-        walletClient
-      }
-    })
-  }
-
-  async function handleFundReward() {
-    let val = Number(amount)
-    if (val === 0 || !account || !publicClient || !walletClient) return
-    val = val * (10 ** token.decimals)
-
-    if (chain?.id !== Number(chainId)) {
-      try {
-        await switchNetworkAsync?.(Number(chainId));
-      } catch (error) {
-        return;
-      }
-    }
-
-    fundReward({
-      gauge,
-      rewardToken: token.address,
-      amount: val,
-      account,
-      clients: {
-        publicClient,
-        walletClient
-      },
-      tokensAtom: [tokens, setTokens]
-    })
-  }
-
-  return (
-    <tr>
-      <td className="!pl-0">
-        <nav className="flex items-center gap-2">
-          <TokenIcon
-            imageSize="w-9 h-9 border border-customNeutral100 rounded-full"
-            token={token}
-            icon={token.logoURI}
-            chainId={chainId}
-          />
-          <strong>${token?.symbol ?? "TKN"}</strong>
-        </nav>
-      </td>
-      <td className="text-left">{formatAndRoundNumber(reward.rate, token.decimals)} {token?.symbol!}/s</td>
-      <td className="text-left">{formatAndRoundNumber(reward.remainingRewards, token.decimals)} {token?.symbol!}</td>
-      <td className="text-left">
-        {reward.periodFinish.toLocaleDateString()}
-      </td>
-      <td className="text-left hidden lg:table-cell">
-        {reward.distributor}
-      </td>
-      <td className="text-left hidden lg:table-cell">
-        <div className="flex flex-row">
-          <div>
-            <InputTokenWithError
-              onSelectToken={() => { }}
-              onMaxClick={handleMaxClick}
-              chainId={chainId}
-              value={amount}
-              onChange={handleChangeInput}
-              selectedToken={token}
-              errorMessage={""}
-              tokenList={[]}
-              allowSelection={false}
-              allowInput={true}
-            />
-          </div>
-          <div className="ml-4">
-            <SecondaryActionButton label="Approve" handleClick={handleApprove} disabled={!account || account !== reward.distributor} className="h-14" />
-            <MainActionButton label="Fund Rewards" handleClick={handleFundReward} disabled={!account || account !== reward.distributor} className="h-14" />
-          </div>
-        </div>
-      </td>
-    </tr>
-  )
 }
